add python3

+++ include/install-nginx.sh
@@ -14,10 +14,15 @@ function install_nginx(){
     
     version="1.8.0"
 
-    nginx_cache_filename="${QUICK_ENV_HOME}/cache/nginx-${version}.tar.gz"
-    nginx_temp_dir="${QUICK_ENV_HOME}/temp/nginx-${version}"
-    nginx_install_dir="${QUICK_ENV_HOME}/local/nginx-${version}"
+    nginx_cache_filename="${QUICK_ENV_CACHE}/nginx-${version}.tar.gz"
+    nginx_source_dir="${QUICK_ENV_CACHE}/nginx-${version}"
+    nginx_install_dir="${QUICK_ENV_LOCAL}/nginx"
     nginx_download_url="http://nginx.org/download/nginx-${version}.tar.gz"
+
+    if [ -e  $nginx_install_dir ]; then
+        echo "nginx install dir exists: ${nginx_install_dir}"
+        return 0
+    fi
 
     # install nginx
     # 1、安装编译工具
@@ -33,9 +38,17 @@ function install_nginx(){
         wget $nginx_download_url -O $nginx_cache_filename
     fi
 
-    tar -zxf $nginx_cache_filename -C "${QUICK_ENV_TEMP}"
-    
-    cd $nginx_temp_dir
+    # 解压
+    if [ -e $nginx_cache_filename ]; then
+        tar -zxf $nginx_cache_filename
+    else
+        echo "nginx cache file not exists!"
+        exit 1
+    fi
+
+    cd $nginx_source_dir
+
+    echo 'nginx installing'
 
     ./configure \
     --prefix=${nginx_install_dir} \
@@ -46,7 +59,8 @@ function install_nginx(){
     --with-http_gzip_static_module \
     --with-pcre \
     --with-http_realip_module \
-    --with-http_sub_module
+    --with-http_sub_module \
+    --quiet
 
     make && make install
 
@@ -55,15 +69,18 @@ function install_nginx(){
 
     # 软链
     # ln -s "${QUICK_ENV_HOME}/local/nginx/1.8.0/sbin/nginx" /usr/local/bin/nginx
-    cat > /etc/profile.d/nginx.sh <<EOF
+    if [ ! -e '/etc/profile.d/nginx.sh' ]; then
+        cat > /etc/profile.d/nginx.sh <<EOF
 # nginx
 export NGINX_HOME=${nginx_install_dir}
 export PATH="\$NGINX_HOME/sbin:\$PATH"
 EOF
+    fi
+
     # 开机自启
     # cp "${QUICK_ENV_HOME}/init.d/nginx.service" /usr/lib/systemd/system/nginx.service
-    
-    cat > /usr/lib/systemd/system/nginx.service <<EOF
+    if [ ! -e '/usr/lib/systemd/system/nginx.service' ]; then
+        cat > /usr/lib/systemd/system/nginx.service <<EOF
 [Unit]
 Description=nginx - high performance web server
 Documentation=http://nginx.org/en/docs/
@@ -84,17 +101,24 @@ LimitCORE=1000000
 [Install]
 WantedBy=multi-user.target
 EOF
-    
+    fi
+
     # 修改配置文件
     mv "${nginx_install_dir}/conf/nginx.conf" "${nginx_install_dir}/conf/nginx.conf.bak"
     cp "${QUICK_ENV_CONFIG}/nginx.conf" "${nginx_install_dir}/conf/nginx.conf"
     
-    mkdir "${nginx_install_dir}/conf/vhost"
+    mkdir -p "${nginx_install_dir}/conf/vhost"
 
     systemctl enable nginx
     systemctl start nginx
 
     # 查看版本
-    source /etc/profile.d/nginx.sh
+    . /etc/profile.d/nginx.sh
     nginx -v
+
+    echo '*************************************'
+    echo '* nginx install success'
+    echo '* source /etc/profile.d/nginx.sh'
+    echo '* systemctl status nginx'
+    echo '*************************************'
 }
